src/main/java/com/aaron/springboot/web/PostsApiController.java
package com.aaron.springboot.web;

import com.aaron.springboot.service.PostsService;
import com.aaron.springboot.web.dto.PostsResponseDto;
import com.aaron.springboot.web.dto.PostsSaveRequestDto;
import com.aaron.springboot.web.dto.PostsUpdateRequestDto;
import lombok.RequiredArgsConstructor;
import org.springframework.web.bind.annotation.*;

/* 스프링에서 Bean을 주입단는 방식은 3가지 이다.
 * 1. @Autowired 2. setter 3. 생성자
 * 이 중 가장 권장하는 방식은 생성자로 주입받는 것이다. */
@RequiredArgsConstructor // 초기화 되지않은 final 필드나, @NonNull 이 붙은 필드에 대해 생성자를 생성해 준다.
// 새로운 필드를 추가할 때 다시 생성자를 만들어서 관리해야하는 번거로움을 없애준다. (@Autowired 를 사용하지 않고 의존성 주입)
@RestController
public class PostsApiController {

    private final PostsService postsService;

    @PostMapping("/api/v1/posts")
    public Long save(@RequestBody PostsSaveRequestDto requestDto) {
        return postsService.save(requestDto);
    }

    @PutMapping("/api/v1/posts/{id}")
    public Long update(@PathVariable Long id, @RequestBody PostsUpdateRequestDto requestDto) {
        return postsService.update(id, requestDto);
    }

    @GetMapping("/api/v1/posts/{id}")
    public PostsResponseDto findById (@PathVariable Long id) {
        return postsService.findById(id);
    }

    @DeleteMapping("/api/v1/posts/{id}")
    public Long delete(@PathVariable Long id) {
        postsService.delete(id);
        return id;
    }
}
